* Add more rounding in SpinBox.qml, allowing FloatPreference::step to work reliably for the spinners without needing double precision (suggestion from ctrlaltdavid: https://github.com/vircadia/vircadia/pull/930#issuecomment-765838930). * Change FloatPreference::step from double back to float.

+++ libraries/shared/src/Preferences.h
@@ -152,7 +152,7 @@ class FloatPreference : public Preference {
     Q_PROPERTY(float value READ getValue WRITE setValue NOTIFY valueChanged)
     Q_PROPERTY(float min READ getMin CONSTANT)
     Q_PROPERTY(float max READ getMax CONSTANT)
-    Q_PROPERTY(double step READ getStep CONSTANT)
+    Q_PROPERTY(float step READ getStep CONSTANT)
     Q_PROPERTY(uint decimals READ getDecimals CONSTANT)
 
 public:
@@ -178,11 +178,11 @@ public:
     float getMax() const { return _max; }
     void setMax(float max) { _max = max; };
 
-    double getStep() const { return _step; }
-    void setStep(double step) { _step = step; };
+    float getStep() const { return _step; }
+    void setStep(float step) { _step = step; };
 
-    uint getDecimals() const { return _decimals; }
-    void setDecimals(uint decimals) { _decimals = decimals; };
+    float getDecimals() const { return _decimals; }
+    void setDecimals(float decimals) { _decimals = decimals; };
 
 signals:
     void valueChanged();
@@ -197,7 +197,7 @@ protected:
     uint _decimals { 0 };
     float _min { 0 };
     float _max { 1 };
-    double _step { 0.1 };
+    float _step { 0.1f };
 };
 
 class IntPreference : public Preference {
